- Create custom scoring for matching candidates - Improved extracting text from pdf file using Document AI - Improved uploading contacts, you can now switch to other pages even uploading is in progress - Align all components related to matching candidates - Improved questionnaire page by adding general questions - Edit form for creating jobs

+++ src/utils/utils.js
@@ -39,7 +39,7 @@ export const handleRedirectToLinkedIn = (link) => {
 
 export const convertDateFormat = (dateString) => {
   const date = new Date(dateString);
-  return new Intl.DateTimeFormat('en-US', { month: 'short', day: '2-digit', year: 'numeric' }).format(date);
+  return new Intl.DateTimeFormat('en-US', { month: 'numeric', day: '2-digit', year: 'numeric' }).format(date);
 };
 
 export const formatDate = (dateString) => {
